# openapi.yaml
openapi: 3.1.0
info:
  title: Groceries API
  version: 1.0.0
  description: |
    REST API for the Groceries application. Provides token-based authentication
    suitable for native app clients and future frontend rewrites.

    ## Authentication

    All endpoints except `POST /api/v1/auth/login` require a Bearer token in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained from the login endpoint and expire after 720 hours (30 days).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  # -------------------------------------------------------------------------
  # Schemas
  # -------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          examples:
            - "item not found"

    # --- Auth ----------------------------------------------------------------
...
    User:
      type: object
      required: [id, name, admin]
      properties:
        id:
          type: integer
          examples:
            - 1
        name:
          type: string
          examples:
            - "alice"
        admin:
          type: boolean
          description: Whether this user has administrator privileges
...
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          examples:
            - "alice"
        admin:
          type: boolean

    # --- Group ---------------------------------------------------------------
    Group:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          examples:
            - 1
        name:
          type: string
          minLength: 3
          examples:
            - "Household"
...
    UpdateGroupRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 3
          examples:
            - "Household"

    # --- Store ---------------------------------------------------------------
...
    Category:
      type: object
      required: [id, store_id, name, description, item_count]
      properties:
        id:
          type: integer
          examples:
            - 1
        store_id:
          type: integer
          examples:
            - 1
        name:
          type: string
          minLength: 3
          examples:
            - "Produce"
        description:
          type: string
          examples:
            - "Fresh fruits and vegetables"
        item_count:
          type: integer
          description: Number of items assigned to this category
          examples:
            - 12
...
    UpdateCategoryRequest:
      type: object
      required: [store_id, name]
      properties:
        store_id:
          type: integer
          examples:
            - 1
        name:
          type: string
          minLength: 3
          examples:
            - "Produce"
        description:
          type: string

    # --- Item ----------------------------------------------------------------
    ListItemSummary:
      type: object
      description: Presence of this object on an Item indicates the item is on the shopping list
      required: [id, quantity, done]
      properties:
        id:
          type: integer
          description: ID of the list entry (not the item itself)
          examples:
            - 42
        quantity:
          type: string
          examples:
            - "2 lbs"
        done:
          type: boolean
          description: Whether this item has been picked up during the current shopping trip
    Item:
      type: object
      required: [id, category_id, category_name, name]
      properties:
        id:
          type: integer
          examples:
            - 1
        category_id:
          type: integer
          examples:
            - 3
        category_name:
          type: string
          examples:
            - "Produce"
        name:
          type: string
          examples:
            - "Apples"
        list:
          $ref: "#/components/schemas/ListItemSummary"
          description: Present when this item is currently on the shopping list
...
    UpdateItemRequest:
      type: object
      required: [category_id, name]
      properties:
        category_id:
          type: integer
          examples:
            - 3
        name:
          type: string
          minLength: 1
          examples:
            - "Apples"

    # --- Shopping list -------------------------------------------------------
...
  responses:
    NoContent:
      description: Operation succeeded with no response body
    BadRequest:
      description: The request was malformed or failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The operation conflicts with existing data (e.g. duplicate name)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerError:
      description: An unexpected server-side error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Numeric resource ID
...
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getUser
      summary: Get a user by ID
      tags: [users]
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateUser
      summary: Update a user
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteUser
      summary: Delete a user
      tags: [users]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Groups
  # --------------------------------------------------------------------------
...
  /api/v1/groups/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getGroup
      summary: Get a group by ID
      tags: [groups]
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateGroup
      summary: Update a group
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteGroup
      summary: Delete a group
      description: Fails if any users are still assigned to the group.
      tags: [groups]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Stores
  # --------------------------------------------------------------------------
...
  /api/v1/categories/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getCategory
      summary: Get a category by ID, including its items
      tags: [categories]
      responses:
        "200":
          description: Category with items
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Category"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateCategory
      summary: Update a category
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteCategory
      summary: Delete a category
      description: Fails if any items are still assigned to the category.
      tags: [categories]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Items
  # --------------------------------------------------------------------------
...
  /api/v1/items/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"

    get:
      operationId: getItem
      summary: Get an item by ID
      tags: [items]
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    put:
      operationId: updateItem
      summary: Update an item
      tags: [items]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      operationId: deleteItem
      summary: Delete an item
      description: Fails if the item is used in any recipes.
      tags: [items]
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --------------------------------------------------------------------------
  # Shopping list
  # --------------------------------------------------------------------------
...